# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FaceGuard - Face Recognition Access & Attendance System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, all]
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Register a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /employees/{id}/photo:
    post:
      operationId: uploadEmployeePhoto
      tags: [employees]
      summary: Upload employee face photo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoUpload"
      responses:
        "200":
          description: Photo uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"


  # CAMERAS
  /cameras:
    get:
      operationId: listCameras
      tags: [cameras]
      summary: List all cameras
      responses:
        "200":
          description: List of cameras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Camera"
    post:
      operationId: createCamera
      tags: [cameras]
      summary: Add a new camera
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CameraInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camera"
  /cameras/{id}:
    get:
      operationId: getCamera
      tags: [cameras]
      summary: Get camera by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Camera
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camera"
    patch:
      operationId: updateCamera
      tags: [cameras]
      summary: Update camera
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CameraUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camera"
    delete:
      operationId: deleteCamera
      tags: [cameras]
      summary: Delete camera
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /cameras/{id}/test:
    post:
      operationId: testCameraConnection
      tags: [cameras]
      summary: Test camera RTSP/HTTP connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CameraTestResult"

  # ZONES
...
  /recognitions:
    get:
      operationId: listRecognitions
      tags: [recognitions]
      summary: List recognition events
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: cameraId
          in: query
          schema:
            type: ["integer", "null"]
        - name: zoneId
          in: query
          schema:
            type: ["integer", "null"]
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: status
          in: query
          schema:
            type: string
            enum: [recognized, unknown, denied, all]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Recognition events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecognitionEvent"
    post:
      operationId: createRecognition
      tags: [recognitions]
      summary: Log a recognition event (called by camera integration)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecognitionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecognitionEvent"

  # ATTENDANCE
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, firstName, lastName, employeeNumber, department, position, status, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        employeeNumber:
          type: string
        department:
          type: string
        position:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        hiredAt:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
    EmployeeInput:
      type: object
      required: [firstName, lastName, employeeNumber, department, position]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        employeeNumber:
          type: string
          minLength: 1
        department:
          type: string
        position:
          type: string
        email:
          type: string
        phone:
          type: string
        hiredAt:
          type: string
          format: date
    EmployeeUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        department:
          type: string
        position:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        hiredAt:
          type: ["string", "null"]
          format: date
    PhotoUpload:
      type: object
      required: [photoBase64]
      properties:
        photoBase64:
          type: string
          description: Base64 encoded image
    Camera:
      type: object
      required: [id, name, brand, protocol, host, zoneId, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
          enum: [dahua, hikvision, unv, other]
        protocol:
          type: string
          enum: [rtsp, http, https]
        host:
          type: string
        port:
          type: ["integer", "null"]
        username:
          type: ["string", "null"]
        streamPath:
          type: ["string", "null"]
        zoneId:
          type: integer
        zoneName:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, offline, unknown]
        createdAt:
          type: string
          format: date-time
    CameraInput:
      type: object
      required: [name, brand, protocol, host, zoneId]
      properties:
        name:
          type: string
          minLength: 1
        brand:
          type: string
          enum: [dahua, hikvision, unv, other]
        protocol:
          type: string
          enum: [rtsp, http, https]
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        password:
          type: string
        streamPath:
          type: string
        zoneId:
          type: integer
    CameraUpdate:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
          enum: [dahua, hikvision, unv, other]
        protocol:
          type: string
          enum: [rtsp, http, https]
        host:
          type: string
        port:
          type: ["integer", "null"]
        username:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        streamPath:
          type: ["string", "null"]
        zoneId:
          type: integer
        status:
          type: string
          enum: [online, offline, unknown]
    CameraTestResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        latencyMs:
          type: ["integer", "null"]
...
    RecognitionEvent:
      type: object
      required: [id, cameraId, status, confidence, detectedAt]
      properties:
        id:
          type: integer
        cameraId:
          type: integer
        cameraName:
          type: ["string", "null"]
        zoneId:
          type: ["integer", "null"]
        zoneName:
          type: ["string", "null"]
        employeeId:
          type: ["integer", "null"]
        employeeName:
          type: ["string", "null"]
        employeeNumber:
          type: ["string", "null"]
        employeePhotoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [recognized, unknown, denied]
        confidence:
          type: number
          description: Confidence score 0-1
        snapshotUrl:
          type: ["string", "null"]
        detectedAt:
          type: string
          format: date-time
    RecognitionInput:
      type: object
      required: [cameraId, status, confidence, detectedAt]
      properties:
        cameraId:
          type: integer
        employeeId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [recognized, unknown, denied]
        confidence:
          type: number
        snapshotUrl:
          type: ["string", "null"]
        detectedAt:
          type: string
          format: date-time
...