# lib/api-zod/src/generated/types/cameraTestResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FaceGuard - Face Recognition Access & Attendance System
 * OpenAPI spec version: 0.1.0
 */

export interface CameraTestResult {
  success: boolean;
  message: string;
  /** @nullable */
  latencyMs?: number | null;
}