# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FaceGuard - Face Recognition Access & Attendance System
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeStatus } from './employeeStatus';

export interface Employee {
  id: number;
  firstName: string;
  lastName: string;
  employeeNumber: string;
  department: string;
  position: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  status: EmployeeStatus;
  /** @nullable */
  hiredAt?: Date | null;
  createdAt: Date;
}