# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FaceGuard - Face Recognition Access & Attendance System
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 1 */
  employeeNumber: string;
  department: string;
  position: string;
  email?: string;
  phone?: string;
  hiredAt?: Date;
}