# lib/api-zod/src/generated/types/recognitionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FaceGuard - Face Recognition Access & Attendance System
 * OpenAPI spec version: 0.1.0
 */
import type { RecognitionInputStatus } from './recognitionInputStatus';

export interface RecognitionInput {
  cameraId: number;
  /** @nullable */
  employeeId?: number | null;
  status: RecognitionInputStatus;
  confidence: number;
  /** @nullable */
  snapshotUrl?: string | null;
  detectedAt: Date;
}